refactor(PublicShareApi): Create a new api repository specific to publicShare
Use non blocking apiCall in the same time
